Move entities to Konveyor.Core.Models namespace
